\name{mypackage-internal}
\alias{Change.Points.R}
\alias{Chisq.Contrib.C}
\alias{Chisq.Contrib.fromS.R}
\alias{Chisq.Contrib.R}
\alias{Chisq.Contrib.test}
\alias{compute.max.ProjectedZ}
\alias{compute.max.Z.C}
\alias{compute.max.Z}
\alias{computeBeta}
\alias{ComputeBYZ.fromS.R.partial}
\alias{ComputeBYZ.fromS.R}
\alias{computeMoments}
\alias{ComputeProjectedZ.fromS.R.partial}
\alias{ComputeProjectedZ.fromS.R}
\alias{computeTiltDirect}
\alias{ComputeZ.C}
\alias{ComputeZ.fromS.R.partial}
\alias{ComputeZ.fromS.R}
\alias{computeZ.onechange}
\alias{computeZ.onechange.sample}
\alias{ComputeZ.R}
\alias{computeZ.squarewave.sample}
\alias{computeZ.test}
\alias{dchi}
\alias{fcompute.max.ProjectedZ}
\alias{fcompute.max.Z}
\alias{flatten}
\alias{fscan.max}
\alias{getCutoffEpidemicChangeLocationKnown}
\alias{getCutoffMultisampleWeightedChisq}
\alias{intmode}
\alias{matrix.max}
\alias{pmarg.sumweightedchisq}
\alias{pvalueMultisampleWeightedChisq}
\alias{pvalueMultisampleWeightedChisqold}
\alias{pvalueSumChisq}
\alias{Remove.Overlap.C}
\alias{Remove.Overlap.R}
\alias{Remove.Overlap.test}
\alias{update.bic.baseline}
\alias{vu}
\alias{wpca}

\title{Internal mpcbs Functions}
\description{
    Internal mpcbs functions
}
\details{
    These are not to be called by the user.
}
\keyword{ internal }
        
