\name{mpcbsexample}
          \docType{data}
          \alias{mpcbsexample}
          \title{Example copy number log ratio data from Illumina, Affymetrix, and Agilent platforms.}
          \description{
            This data set contains an example copy number data set where the same genomic region is assayed
            using the Illumina, Affymetrix, and Agilent platforms. 
          }
          \usage{mpcbsexample}
          \format{Contains the fields "illu", "affy", "agil" each of which is a two column matrix where the first column is 
          genome position (in base pairs) and the second column is a log ratio.
                
          }
          \source{TCGA data.}
          \references{
            The TCGA Consortium.
          }
          \keyword{datasets}
